requirements.txt
streamlit
pandas
Pillow
pytest
ruff